Extract pure feed/release parsers; dedupe station-WiFi setup
RssFeedManager and OtaUpdater each mixed three concerns: WiFi/HTTP
transport, response parsing, and feature orchestration. The parsing
could only run on the device, and the WiFi connect/poll/timeout loop
was duplicated byte-for-byte between the two managers.

Pull the pure logic out behind small interfaces:

- rss/FeedParser     RSS/Atom item parsing + HTML stripping, XML entity
                     decoding, character folding. String in, FeedItem out.
- update/ReleaseParser  GitHub releases/latest JSON -> tag + asset URL.
                     String in, ReleaseInfo out.
- net/WifiConnection  the single WIFI_STA connect/poll/timeout loop both
                     managers shared; timeout policy now lives in one place.

The managers keep their public interfaces and behavior and delegate to
these modules, shrinking to transport + orchestration. The RSS connect
progress ramp is unified with OTA's (cosmetic, progress bar only).

Tests: both parsers get host-side unit tests, joining ReadingLoop as
natively-tested modules; the Arduino String shim was extended to support
them. 73 native test cases pass; both ESP32 envs build.
