+++ platformio.ini
@@ -30,9 +30,16 @@ build_unflags =
 platform = native
 framework =
 test_framework = unity
-test_filter = test_pacing
+test_filter =
+  test_pacing
+  test_release_parser
+  test_feed_parser
 test_build_src = yes
-build_src_filter = -<*> +<reader/ReadingLoop.cpp>
+build_src_filter =
+  -<*>
+  +<reader/ReadingLoop.cpp>
+  +<update/ReleaseParser.cpp>
+  +<rss/FeedParser.cpp>
 build_flags =
   -std=c++17
   -Itest/support
