+++ src/rss/RssFeedManager.h
@@ -3,6 +3,7 @@
 #include <Arduino.h>
 #include <Preferences.h>
 
+#include "rss/FeedParser.h"
 #include "update/OtaUpdater.h"
 
 class RssFeedManager {
@@ -21,13 +22,6 @@ class RssFeedManager {
                     StatusCallback callback = nullptr, void *context = nullptr);
 
  private:
-  struct FeedItem {
-    String title;
-    String link;
-    String author;
-    String body;
-  };
-
   bool connectWiFi(const OtaUpdater::Config &wifiConfig, StatusCallback callback, void *context);
   void disconnectWiFi();
   bool fetchUrl(const String &url, String &body, String &error, uint8_t feedIndex,
@@ -35,21 +29,12 @@ class RssFeedManager {
   bool processFeed(const String &feedUrl, const String &feedBody, Preferences &preferences,
                    Result &result, uint8_t feedIndex, uint8_t feedCount, StatusCallback callback,
                    void *context);
-  bool parseNextItem(const String &feedBody, size_t &searchStart, FeedItem &item);
-  bool saveItem(const FeedItem &item, Preferences &preferences, Result &result);
-  bool itemAlreadySeen(const FeedItem &item, Preferences &preferences);
-  void markItemSeen(const FeedItem &item, Preferences &preferences);
-  String seenKeyForItem(const FeedItem &item) const;
-  String itemIdentity(const FeedItem &item) const;
-  String valueBetween(const String &text, const String &openTag, const String &closeTag,
-                      size_t start, size_t end) const;
-  String attributeValue(const String &text, const String &tagPrefix, const String &attribute,
-                        size_t start, size_t end) const;
-  String cleanText(String value) const;
-  String stripHtml(const String &html) const;
-  String xmlDecode(String value) const;
-  String sourceLabelForItem(const FeedItem &item) const;
-  String filenameForItem(const FeedItem &item) const;
+  bool saveItem(const feedparser::FeedItem &item, Preferences &preferences, Result &result);
+  bool itemAlreadySeen(const feedparser::FeedItem &item, Preferences &preferences);
+  void markItemSeen(const feedparser::FeedItem &item, Preferences &preferences);
+  String seenKeyForItem(const feedparser::FeedItem &item) const;
+  String itemIdentity(const feedparser::FeedItem &item) const;
+  String filenameForItem(const feedparser::FeedItem &item) const;
   String metadataSafe(String value) const;
   uint32_t fnv1a(const String &value) const;
   void report(StatusCallback callback, void *context, const String &line1, const String &line2,
